components: extract stat_card as separate component
